requirements.txt
alembic==1.13.1
asyncpg==0.29.0
psycopg==3.1.18
psycopg-binary==3.1.10
sqlalchemy==2.0.29

fastapi==0.111.0
pydantic==2.7.1
pydantic-settings==2.2.1
unicorn==2.0.1
python-jose=3.3.0
